Added bottle logo UV template, and texture folder
added a texture folder in the art folder with a Krita document, and a UV template for all the bottle logos. I also included a sample that shows all the meshes of the bottles. The order of the meshes in the sample image is not the same order of the UVs on the texture template by the way.